+++ src/global/tracker_layout_loader/tracker_layout_loader.gd
@@ -82,8 +82,8 @@ func parse_item_params(params: Dictionary) -> ItemHint:
 				if value is String:
 					result.description = value.strip_edges()
 			"icons":
-				if value is float:
-					result.icons = int(value)
+				if value is Array:
+					result.icons.assign(value)
 			"is_pinned":
 				if value is bool:
 					result.pinned = value